requirements.txt
Django==1.9.8
django-jquery==3.1.0
django-static-jquery==2.1.4
djangorestframework==3.4.1
